requirements.txt
asgiref
attrs
colorama
Django
djangorestframework
drf-spectacular
inflection
jsonschema
jsonschema-specifications
pillow
python-dotenv
PyYAML
qrcode
referencing
rpds-py
sqlparse
tzdata
uritemplate
